refactor($CarName, RacingGame): delete attempt validator and mapper
