+++ src/main/java/racingcar/model/vo/CarName.java
@@ -7,7 +7,29 @@ public class CarName {
     private final String name;
 
     public CarName(String carName) {
+        validate(carName);
         this.name = carName;
+    }
+
+    public static CarName from(String carName) {
+        return new CarName(carName);
+    }
+
+    private void validate(String carName) {
+        isLengthLessThan5(carName);
+        isBlank(carName);
+    }
+
+    private void isLengthLessThan5(String carName) {
+        if (carName.length() >= 5) {
+            throw new IllegalArgumentException("이름은 5자 이내여야 합니다.");
+        }
+    }
+
+    public void isBlank(String carName) {
+        if (carName.isBlank()) {
+            throw new IllegalArgumentException("이름에 빈 값이 들어올 수 없습니다.");
+        }
     }
 
     @Override
